parser/requirements.txt
langchain-core
langchain-openai
python-dotenv
PyPDF2